requirements.txt
streamlit==1.37.1
python-dotenv==1.0.1
openai==1.11.0
torch==2.2.0
torchvision==0.17.0
Pillow==10.0.0
numpy==1.26.0
sqlalchemy==2.2.5
pandas==2.2.2
